requirements.txt
playwright==1.44.0
aiohttp==3.9.5
aiofiles==23.2.1
pyyaml==6.0.1
python-dotenv==1.0.1
loguru==0.7.2
asyncio-throttle==1.0.2
fake-useragent==1.5.1
tenacity==8.3.0
tabulate==0.9.0
